Player, gameboard, ship factories created
Player imports gameboard and ship factories. Player can check
whether a ship can be placed in the X or Y orientation based on the
length of the ship in question. If the spaces on the board are 0, the
ship can be placed. If they are undefined (out of bounds) or occupied
by another ship object, the function will return false.